acpi: move stream reader implementation into parser pkg
