+++ src/gopheros/device/acpi/aml/parser/stream_reader_test.go
@@ -1,4 +1,4 @@
-package aml
+package parser
 
 import (
 	"io"
